% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{after_ortho_format}
\alias{after_ortho_format}
\title{Formatted output of OrthoMCL.}
\format{List of 2: (1) presence absence matrix, (2) protein ids:
\describe{
  \item{pa_matrix}{matrix showing taxa presence/absence in OG}
  \item{proteins}{matrix listing protein_id contained in each OG}
}}
\usage{
after_ortho_format
}
\description{
A list created by inputting the output of OrthoMCL clusters
into the format_afterOrtho function.
}
\keyword{datasets}
