% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePrincipalCoordinates.R
\name{CalculatePrincipalCoordinates}
\alias{CalculatePrincipalCoordinates}
\title{Show Principal Components Breakdown}
\usage{
CalculatePrincipalCoordinates(mcl_data)
}
\arguments{
\item{mcl_data}{output of FormatAfterOrtho --list of 2 things-- 1: binary matrix indicating the presence / absence of genes in each OG and 2: vector of names of OGs}
}
\value{
returns a named list of principal components and accompanying proportion of variance for each
}
\description{
Function to show Principal Components statistics based on the OrthoMCL presence absence groupings.
}
\examples{
CalculatePrincipalCoordinates(after_ortho_format)
}
