% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_repset.R
\name{join_repset}
\alias{join_repset}
\title{Join Representative Sequences}
\usage{
join_repset(reps_file, mcl_mtrx)
}
\arguments{
\item{reps_file}{output of pick_repseq()}

\item{mcl_mtrx}{OrthoMCL output matrix from analyze_OrthoMCL()}
}
\value{
Returns the original OrthoMCL output matrix with additional columns: representative sequence taxon, representative sequence id, representative sequence annotation, representative sequence
}
\description{
Joins the OrthoMCL output matrix to representative sequences
}
\examples{
joined_mtrx_grps <- join_repset(repseqs_grps, mcl_mtrx_grps)
#mcl_mtrx previously derived from analyze_OrthoMCL()
#repseqs previously derived from pick_repseq()
}

