% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{joined_mtrx_grps}
\alias{joined_mtrx_grps}
\title{Final output of join_repset.}
\format{A data frame with 10 rows and 11 variables:
\describe{
  \item{COG}{taxa cluster id, as defined by OrthoMCL}
  \item{pval1}{p-value, based on presence absence}
  \item{corrected_pval1}{Bonferroni p-value, corrected by number of tests}
  \item{mean_COGContain}{mean of all taxa phenotypes in that COG}
  \item{mean_COGLack}{mean of all taxa phenotypes not in that COG}
  \item{taxa_contain}{taxa in that cluster}
  \item{taxa_miss}{taxa not in that cluster}
  \item{rep_taxon}{randomly selected representative taxa from the cluster}
  \item{rep_id}{protein id, from randomly selected representative taxa}
  \item{rep_annot}{fasta annotation, from randomly selected representative taxa}
  \item{rep_seq}{AA sequence, from randomly selected representative taxa}
}}
\usage{
joined_mtrx_grps
}
\description{
A data frame containing the final results of statistical analysis with 
protein ids, annotations, and sequences added.
}
\keyword{datasets}

