% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{manhat_grp}
\alias{manhat_grp}
\title{Manhattan Plot of All Taxa}
\usage{
manhat_grp(mcl_data, mcl_mtrx, tree = NULL)
}
\arguments{
\item{mcl_data}{format_afterOrtho output}

\item{mcl_mtrx}{output of analyze_OrthoMCL()}

\item{tree}{tree file optional, used for ordering taxa along x axis}
}
\value{
a manhattan plot
}
\description{
Manhattan plot that graphs all p-values for taxa.
}
\examples{
manhat_grp(after_ortho_format, mcl_mtrx)

#@param equation of line of significance, defaults to -log10((.05)/dim(pdgs)[1])
}
\references{
Some sort of reference
}
