% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{mcl_mtrx}
\alias{mcl_mtrx}
\title{Final output of AnalyzeOrthoMCL}
\format{A matrix with 17 rows and 7 variables:
\describe{
  \item{OG}{taxa cluster id, as defined by OrthoMCL}
  \item{pval1}{p-value, based on presence absence}
  \item{corrected_pval1}{Bonferroni p-value, corrected by number of tests}
  \item{mean_OGContain}{mean of all taxa phenotypes in that OG}
  \item{mean_OGLack}{mean of all taxa phenotypes not in that OG}
  \item{taxa_contain}{taxa in that cluster}
  \item{taxa_miss}{taxa not in that cluster}
}}
\usage{
mcl_mtrx
}
\description{
A matrix containing the final results of statistical analysis.
}
\keyword{datasets}
