% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pdg_v_cog}
\alias{pdg_v_cog}
\title{Number of PDGs vs COGs/PDG}
\usage{
pdg_v_cog(mcl_data, num = 40, ...)
}
\arguments{
\item{mcl_data}{format_afterOrtho output}

\item{num}{an integer indicating where the x axis should end and be compiled}

\item{...}{args to be passed to barplot}
}
\value{
a barplot with a height determined by the second column and the first column abbreviated to accomodate visual spacing
}
\description{
Barplot that indicates the number of PDGs vs COGs(clustered orthologous groups) in a PDG
}
\examples{
pdg_v_cog(after_ortho_format_grps,2)
#dev.off() #reset margins
}
\references{
Some sort of reference
}

