% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pdg_v_OG}
\alias{pdg_v_OG}
\title{Number of PDGs vs OGs/PDG}
\usage{
pdg_v_OG(mcl_data, num = 40, ...)
}
\arguments{
\item{mcl_data}{format_afterOrtho output}

\item{num}{an integer indicating where the x axis should end and be compiled}

\item{...}{args to be passed to barplot}
}
\value{
a barplot with a height determined by the second column and the first column abbreviated to accomodate visual spacing
}
\description{
Barplot that indicates the number of PDGs vs OGs(clustered orthologous groups) in a PDG
}
\examples{
# Not run - changes options
\dontrun{
pdg_v_OG(after_ortho_format_grps,2)
}
}
