% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{PDGvOG}
\alias{PDGvOG}
\title{Number of PDGs vs OGs/PDG}
\usage{
PDGvOG(mcl_data, num = 40, ...)
}
\arguments{
\item{mcl_data}{FormatAfterOrtho output}

\item{num}{an integer indicating where the x axis should end and be compiled}

\item{...}{args to be passed to barplot}
}
\value{
a barplot with a height determined by the second column and the first column abbreviated to accomodate visual spacing
}
\description{
Barplot that indicates the number of PDGs vs OGs(clustered orthologous groups) in a PDG
}
\examples{

PDGvOG(after_ortho_format_grps,2)

}
