% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatting.R
\name{pick_repseq}
\alias{pick_repseq}
\title{Pick Representative Sequences}
\usage{
pick_repseq(mcl_file, fa_dir, del_mid = T, format = "ortho")
}
\arguments{
\item{mcl_file}{OrthoMCL output file}

\item{fa_dir}{Path to the directory where all raw GenBank files are stored. Note, all file names must be changed to a 4-letter code representing each species and have ".fasta" file descriptor}

\item{del_mid}{Print intermediate fully concatinated fasta (default: FALSE)}

\item{format}{Select format of OrthoMCL output file (default: "ortho" for version 2.0 orthologGroups, option: "groups" for version 1.0 output)}
}
\value{
The path to the representative sequence file in fasta format
}
\description{
Randomly picks a representative sequence from GenBank fasta files for every OrthoMCL COG
}
\examples{
repseqfile <- pick_repseq("small_data/orthologGroups", "small_data")
}

