% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{QQPlotter}
\alias{QQPlotter}
\title{QQPlot}
\usage{
QQPlotter(mcl_mtrx)
}
\arguments{
\item{mcl_mtrx}{matrix generated by AnalyzeOrthoMCL}
}
\value{
a qqplot of the p-values obtained through AnalyzeOrthoMCL
}
\description{
Makes a qqplot of the p-values obtained through AnalyzeOrthoMCL
}
\examples{
QQPlotter(mcl_mtrx)
}
\references{
Some sore of reference
}
