% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{repseqs}
\alias{repseqs}
\title{Output of pick_repseq.}
\format{A matrix with 34 rows and 1 variables:
\describe{
  \item{NULL}{data to be used in joinrepset}
}}
\usage{
repseqs
}
\description{
An intermediate step in creating the joined matrix.
}
\keyword{datasets}

