% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{starv_pheno_data}
\alias{starv_pheno_data}
\title{Starvation rate of fruit flies dataset.}
\format{A matrix with 543 rows and 7 variables:
\describe{
  \item{EXP}{random effect variable, experiment number of flies}
  \item{VIAL}{random effect variable, vial number of flies}
  \item{BACLO}{fixed effect variable, loss of bacteria in flies}
  \item{TRT}{4-letter taxa designation of associated bacteria}
  \item{t1}{time 1}
  \item{t2}{time 2}
  \item{event}{event}
}}
\usage{
starv_pheno_data
}
\description{
A subset of the Starvation rate of fruit flies, collected in the Chaston Lab,
to be used as a brief example for survival tests in analyzeOrthoMCL.
}
\keyword{datasets}
