% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mcl.R
\name{write_mcl}
\alias{write_mcl}
\title{Print analyzed matrix}
\usage{
write_mcl(mtrx, filename)
}
\arguments{
\item{mtrx}{Matrix derived from analyze_OrthoMCL}

\item{filename}{File name to save final output}
}
\value{
The path to the written file
}
\description{
Writes a tab separated version of the analyzed OrthoMCL data with or without the joined representative sequences
}
\examples{
write_mcl(mcl_mtrx, 'matrix.tsv')
#mcl_mtrx previously derived from analyze_OrthoMCL() or join_repset()
}
