% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteMCL.R
\name{WriteMCL}
\alias{WriteMCL}
\title{Print analyzed matrix}
\usage{
WriteMCL(mtrx, filename)
}
\arguments{
\item{mtrx}{Matrix derived from AnalyzeOrthoMCL}

\item{filename}{File name to save final output}
}
\value{
The path to the written file
}
\description{
Writes a tab separated version of the analyzed OrthoMCL data with or without the joined representative sequences
}
\examples{

\dontrun{
WriteMCL(mcl_mtrx, 'matrix.tsv')
#mcl_mtrx previously derived from AnalyzeOrthoMCL() or join_repset()
}
}
