% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyzeOrthoMCL.R
\name{analyze_OrthoMCL}
\alias{analyze_OrthoMCL}
\title{Main OrthoMCL Analysis}
\usage{
analyze_OrthoMCL(mcl_file, var_file, species_name, model, resp = NULL,
  fix2 = NULL, rndm1 = NULL, rndm2 = NULL, format = "ortho")
}
\arguments{
\item{mcl_file}{Path to OrthoMCL output file containing COG (cluster of orthologous groups) groupings}

\item{var_file}{Path to tab separated phenotype data}

\item{species_name}{Column name in var_file containing 4-letter species designations}

\item{model}{Linear Model with gene presence as fixed effect (lm), Wilcox Test with gene presence as fixed effect (wx), Linear Mixed Effect models with gene presence as fixed effect and additional variables specified as: one random effect (lmeR1); two independent random effects (lmeR2ind); two random effects with rndm2 nested in rndm1 (lmeR2nest); or two independent random effects with one additional fixed effect (lmeF2)}

\item{resp}{Column name in var_file containing response variable}

\item{fix2}{Column name in var_file containing second fixed effect}

\item{rndm1}{Column name in var_file containing first random variable}

\item{rndm2}{Column name in var_file containing second random variable}

\item{format}{Select format of OrthoMCL output file (default: "ortho" for version 2.0 orthologGroups, option: "groups" for version 1.0 output)}
}
\value{
A matrix with the following columns: PDG, p-values, Bonferroni corrected p-values, mean phenotype of PDG-containing taxa, mean pheotype of PDG-lacking taxa, taxa that contain PDG, taxa that lack PDG
}
\description{
Main function for analyzing the statistical association of PDG (phylogenetic distribution group) presence with phenotype data
}
\examples{
#Linear Model
mcl_mtrx <- analyze_OrthoMCL("sample_data/orthologGroups", "sample_data/pheno_data.tsv", "Treatment", "lm", resp="RespVar")

#Wilcox Test
mcl_mtrx <- analyze_OrthoMCL("sample_data/orthologGroups", "sample_data/pheno_data.tsv", "Treatment", "wx", resp="RespVar")

#Linear Mixed Effect with one random effect
mcl_mtrx <- analyze_OrthoMCL("sample_data/orthologGroups", "sample_data/pheno_data.tsv", "Treatment", "lmeR1", resp="RespVar", rndm1="Experiment")

#Linear Mixed Effect with two independent random effects
mcl_mtrx <- analyze_OrthoMCL("sample_data/orthologGroups", "sample_data/pheno_data.tsv", "Treatment", "lmeR2ind", resp="RespVar", rndm1="Experiment", rndm2="Vial")

#Linear Mixed Effect with rndm2 nested in rndm1
mcl_mtrx <- analyze_OrthoMCL("sample_data/orthologGroups", "sample_data/pheno_data.tsv", "Treatment", "lmeR2nest", resp="RespVar", rndm1="Experiment", rndm2="Vial")

#Linear Mixed Effect with two independent random effects and one additional fixed effect
mcl_mtrx <- analyze_OrthoMCL("sample_data/orthologGroups", "sample_data/pheno_data.tsv", "Treatment", "lmeF2", resp="RespVar", fix2="Treatment", rndm1="Experiment", rndm2="Vial")
}
\references{
Some sore of reference
}

