% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatting.R
\name{join_repset}
\alias{join_repset}
\title{Join Representative Sequences}
\usage{
join_repset(reps_file, mcl_mtrx)
}
\arguments{
\item{reps_file}{A fasta file containing one representative sequence for every COG in mcl_mtrx}

\item{mcl_mtrx}{OrthoMCL output matrix from analyze_OrthoMCL()}
}
\value{
Returns the original OrthoMCL output matrix with additional columns: representative sequence taxon, representative sequence id, representative sequence annotation, representative sequence
}
\description{
Joins the OrthoMCL output matrix to representative sequences
}
\examples{
joined_mtrx <- join_repset("small_data/repseq.fasta", mcl_mtrx)
#mcl_mtrx previously derived from analyze_OrthoMCL()
}

