% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCOGseqs.R
\name{printOGseqs}
\alias{printOGseqs}
\title{Print OG Sequences}
\usage{
printOGseqs(after_ortho, OG, fasta_dir, out_dir = NULL, outfile = "none")
}
\arguments{
\item{after_ortho}{output from format_afterOrtho}

\item{OG}{name of OG}

\item{fasta_dir}{directory to fastas}

\item{out_dir}{complete path to output directory}

\item{outfile}{name of file that will be written to}
}
\value{
A fasta file with all protein sequnces and ids for a given OG
}
\description{
Print all protein sequences and annotations in a given OG
}
\examples{

# Not run ~ directory structure depends on system
\dontrun{
OG <- 'OG5_126968'
dir <- system.file('extdata', 'fasta_dir', package='MAGNAMWAR')
dir <- paste(dir,'/',sep='')

printOGseqs(after_ortho_format, OG, dir)
}

}
