% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{qqplotter}
\alias{qqplotter}
\title{QQPlot}
\usage{
qqplotter(mcl_mtrx)
}
\arguments{
\item{mcl_mtrx}{matrix generated by analyze_OrthoMCL}
}
\value{
a qqplot of the p-values obtained through analyze_OrthoMCL
}
\description{
Makes a qqplot of the p-values obtained through analyze_OrthoMCL
}
\examples{
qqplotter(mcl_mtrx)
}
\references{
Some sore of reference
}
